IOS-5589 Fix review comments
